working/requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Themes2==0.1.4
Flask-WTF==0.12
Flask-Mail==0.9.1
requests==2.8.1
PyYAML==3.11
lxml==3.6.0
gunicorn==19.4.5
schedule==0.3.2
futures==3.0.5
